chore(client-adapters): moved to folder
