+++ src/client-adapters/client-adapter.js
