+++ src/client-adapters/fetch-client-adapter.js
