+++ src/client-adapters/file-client-adapter.js
