+++ src/client-adapters/file-client.js
