+++ src/client-adapters/http-client-adapter.js
