+++ src/client-adapters/jsonp-client-adapter.js
