+++ src/client-adapters/storage-client-adapter.js
