+++ src/client-adapters/util.js
