+++ test/client-adapters/http-client-adapter.spec.js
@@ -1,17 +1,16 @@
-import {FetchClientAdapter} from '../src/fetch-client-adapter';
-import {HttpClient} from 'aurelia-fetch-client';
+import {HttpClientAdapter} from '../../src/client-adapters/http-client-adapter';
+import {HttpClient} from 'aurelia-http-client';
 import {buildQueryString} from 'aurelia-path';
-import {settings} from './resources/settings';
+import {settings} from '../resources/settings';
 
-let adapter = new FetchClientAdapter();
+let adapter = new HttpClientAdapter();
 
-describe('FetchClientAdapter', function() {
+describe('HttpClientAdapter', function() {
   describe('.client', function() {
     it('Should be client with configure(config => config.withBaseUrl(base))', function() {
       expect(adapter.client instanceof HttpClient).toBe(true);
 
       adapter.client.configure(config => config.withBaseUrl(settings.baseUrls.api));
-      expect(adapter.client.baseUrl).toBe(settings.baseUrls.api);
     });
   });
 
@@ -78,7 +77,7 @@ describe('FetchClientAdapter', function() {
   });
 
   describe('.destroy()', function() {
-    it('Should destroy with id and options .', function(done) {
+    it('Should destroy with id and options.', function(done) {
       adapter.request('DELETE', 'posts/id', undefined, settings.options)
         .then(y => {
           expect(y.method).toBe('DELETE');
